refactor(arena/tui): split conversation panel rendering from state (#1456)
